refactor: xmodule/seq_module.py -> xmodule/seq_block.py
