+++ lms/djangoapps/course_blocks/transformers/hidden_content.py
@@ -8,7 +8,7 @@ from datetime import datetime
 from pytz import utc
 
 from openedx.core.djangoapps.content.block_structure.transformer import BlockStructureTransformer
-from xmodule.seq_module import SequenceBlock  # lint-amnesty, pylint: disable=wrong-import-order
+from xmodule.seq_block import SequenceBlock  # lint-amnesty, pylint: disable=wrong-import-order
 
 from .utils import collect_merged_boolean_field, collect_merged_date_field
