+++ setup.py
@@ -8,7 +8,7 @@ XBLOCKS = [
     "about = xmodule.html_block:AboutBlock",
     "book = xmodule.template_module:TranslateCustomTagBlock",
     "annotatable = xmodule.annotatable_block:AnnotatableBlock",
-    "chapter = xmodule.seq_module:SectionBlock",
+    "chapter = xmodule.seq_block:SectionBlock",
     "conditional = xmodule.conditional_block:ConditionalBlock",
     "course = xmodule.course_block:CourseBlock",
     "course_info = xmodule.html_block:CourseInfoBlock",
@@ -27,7 +27,7 @@ XBLOCKS = [
     "poll_question = xmodule.poll_block:PollBlock",
     "problem = xmodule.capa_block:ProblemBlock",
     "randomize = xmodule.randomize_block:RandomizeBlock",
-    "sequential = xmodule.seq_module:SequenceBlock",
+    "sequential = xmodule.seq_block:SequenceBlock",
     "slides = xmodule.template_module:TranslateCustomTagBlock",
     "split_test = xmodule.split_test_module:SplitTestBlock",
     "static_tab = xmodule.html_block:StaticTabBlock",
