+++ xmodule/conditional_block.py
@@ -17,7 +17,7 @@ from xblock.fields import ReferenceList, Scope, String
 from openedx.core.djangolib.markup import HTML, Text
 from xmodule.mako_block import MakoTemplateBlockBase
 from xmodule.modulestore.exceptions import ItemNotFoundError
-from xmodule.seq_module import SequenceMixin
+from xmodule.seq_block import SequenceMixin
 from xmodule.studio_editable import StudioEditableBlock
 from xmodule.util.xmodule_django import add_webpack_to_fragment
 from xmodule.validation import StudioValidation, StudioValidationMessage
