+++ xmodule/course_block.py
@@ -23,7 +23,7 @@ from xmodule import course_metadata_utils
 from xmodule.course_metadata_utils import DEFAULT_GRADING_POLICY, DEFAULT_START_DATE
 from xmodule.data import CertificatesDisplayBehaviors
 from xmodule.graders import grader_from_conf
-from xmodule.seq_module import SequenceBlock
+from xmodule.seq_block import SequenceBlock
 from xmodule.tabs import CourseTabList, InvalidTabsException
 
 from .fields import Date
