+++ xmodule/randomize_block.py
@@ -8,7 +8,7 @@ from lxml import etree
 from web_fragments.fragment import Fragment
 from xblock.fields import Integer, Scope
 from xmodule.mako_block import MakoTemplateBlockBase
-from xmodule.seq_module import SequenceMixin
+from xmodule.seq_block import SequenceMixin
 from xmodule.xml_module import XmlMixin
 from xmodule.x_module import (
     HTMLSnippet,
